requirements.txt
Django==4.2.25
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
mysqlclient==2.2.7
psycopg2-binary==2.9.10
Pillow==11.3.0
Faker==37.12.0
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
gunicorn==22.0.0
